fix(fastqcParse): fastqcParser now accounts for missing modules
fastqcParser used to assume you would have data for all 11 modules, however this is not the case for some samples as they may be missing tile quality because tehy are not Illumina data, or they may not have overrepresented sequences. In that case the indexing for modules became erroneous, which triggered issues with column names and data insertion. It is now fixed so that it only parses for modules which it finds in the fastqc_data.txt file.

close #28